remove EF support to make it agnostic of any ORM
