+++ examples/X.PagedList.Mvc.Example.Core/Program.cs
@@ -36,11 +36,10 @@ public class Program
 
         app.UseAuthorization();
 
-        app.UseEndpoints(endpoints =>
-        {
-            endpoints.MapControllerRoute(
+        app.MapControllerRoute(
                 name: "default",
                 pattern: "{controller=Home}/{action=Index}/{id?}");
-        });
+
+        app.Run();
     }
 }
