+++ src/X.PagedList.Mvc.Core/X.PagedList.Mvc.Core.csproj
@@ -18,7 +18,7 @@
     <AssemblyOriginatorKeyFile>xpagedlist.snk</AssemblyOriginatorKeyFile>
     <LangVersion>default</LangVersion>
     <PackageVersion>8.4.3</PackageVersion>
-    <TargetFramework>net6.0</TargetFramework>
+    <TargetFramework>net7.0</TargetFramework>
   </PropertyGroup>
 
   <ItemGroup>
