+++ tests/X.PagedList.Tests/Blog.cs
@@ -1,0 +1,10 @@
+﻿namespace X.PagedList.Tests;
+
+public class Blog
+{
+    public int BlogID { get; set; }
+
+    public string Name { get; set; }
+
+    public string Url { get; set; }
+}
