+++ tests/X.PagedList.Tests/DbContext.cs
@@ -1,0 +1,3 @@
+﻿namespace X.PagedList.Tests;
+
+public class DbContext { }
