+++ tests/X.PagedList.Tests/DbSet.cs
@@ -1,0 +1,5 @@
+﻿using System.Collections.Generic;
+
+namespace X.PagedList.Tests;
+
+public class DbSet<T> : List<T> { }
