+++ tests/X.PagedList.Tests/IDbAsyncEnumerable.cs
@@ -1,0 +1,4 @@
+﻿namespace X.PagedList.Tests;
+
+public interface IDbAsyncEnumerable { IDbAsyncEnumerator GetAsyncEnumerator(); }
+public interface IDbAsyncEnumerable<T> : IDbAsyncEnumerable { }
