+++ tests/X.PagedList.Tests/IDbAsyncEnumerator.cs
@@ -1,0 +1,4 @@
+﻿namespace X.PagedList.Tests;
+
+public interface IDbAsyncEnumerator<T> : IDbAsyncEnumerator { }
+public interface IDbAsyncEnumerator { object Current { get; } }
