+++ tests/X.PagedList.Tests/IDbAsyncQueryProvider.cs
@@ -1,0 +1,5 @@
+﻿using System.Linq;
+
+namespace X.PagedList.Tests;
+
+public interface IDbAsyncQueryProvider : IQueryProvider { }
