+++ tests/X.PagedList.Tests/TestContext.cs
@@ -1,9 +1,4 @@
 ﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Linq.Expressions;
-using System.Threading;
-using System.Threading.Tasks;
 
 namespace X.PagedList.Tests;
 
@@ -11,118 +6,3 @@ public class TestContext : DbContext
 {
     public virtual DbSet<Blog> Blogs { get; set; }
 }
-
-public class Blog
-{
-    public int BlogID { get; set; }
-
-    public string Name { get; set; }
-
-    public string Url { get; set; }
-}
-
-internal class TestDbAsyncQueryProvider<TEntity> : IDbAsyncQueryProvider
-{
-    private readonly IQueryProvider _inner;
-
-    internal TestDbAsyncQueryProvider(IQueryProvider inner)
-    {
-        _inner = inner;
-    }
-
-    public IQueryable CreateQuery(Expression expression)
-    {
-        return new TestDbAsyncEnumerable<TEntity>(expression);
-    }
-
-    public IQueryable<TElement> CreateQuery<TElement>(Expression expression)
-    {
-        return new TestDbAsyncEnumerable<TElement>(expression);
-    }
-
-    public object Execute(Expression expression)
-    {
-        return _inner.Execute(expression);
-    }
-
-    public TResult Execute<TResult>(Expression expression)
-    {
-        return _inner.Execute<TResult>(expression);
-    }
-
-    public Task<object> ExecuteAsync(Expression expression, CancellationToken cancellationToken)
-    {
-        return Task.FromResult(Execute(expression));
-    }
-
-    public Task<TResult> ExecuteAsync<TResult>(Expression expression, CancellationToken cancellationToken)
-    {
-        return Task.FromResult(Execute<TResult>(expression));
-    }
-
-    internal class TestDbAsyncEnumerable<T> : EnumerableQuery<T>, IDbAsyncEnumerable<T>, IQueryable<T>
-    {
-        public TestDbAsyncEnumerable(IEnumerable<T> enumerable)
-            : base(enumerable)
-        { }
-
-        public TestDbAsyncEnumerable(Expression expression)
-            : base(expression)
-        { }
-
-        public IDbAsyncEnumerator<T> GetAsyncEnumerator()
-        {
-            return new TestDbAsyncEnumerator<T>(this.AsEnumerable().GetEnumerator());
-        }
-
-        IDbAsyncEnumerator IDbAsyncEnumerable.GetAsyncEnumerator()
-        {
-            return GetAsyncEnumerator();
-        }
-
-        IQueryProvider IQueryable.Provider
-        {
-            get { return new TestDbAsyncQueryProvider<T>(this); }
-        }
-    }
-
-    internal class TestDbAsyncEnumerator<T> : IDbAsyncEnumerator<T>
-    {
-        private readonly IEnumerator<T> _inner;
-
-        public TestDbAsyncEnumerator(IEnumerator<T> inner)
-        {
-            _inner = inner;
-        }
-
-        public void Dispose()
-        {
-            _inner.Dispose();
-        }
-
-        public Task<bool> MoveNextAsync(CancellationToken cancellationToken)
-        {
-            return Task.FromResult(_inner.MoveNext());
-        }
-
-        public T Current
-        {
-            get { return _inner.Current; }
-        }
-
-        object IDbAsyncEnumerator.Current
-        {
-            get { return Current; }
-        }
-    }
-}
-
-public class DbContext { }
-public class DbSet<T> : List<T> { }
-public class TestDbAsyncQueryProvider { }
-
-public interface IDbAsyncEnumerable<T> : IDbAsyncEnumerable { }
-public interface IDbAsyncEnumerator<T> : IDbAsyncEnumerator { }
-public interface IDbAsyncQueryProvider : IQueryProvider { }
-public interface IDbAsyncEnumerable { IDbAsyncEnumerator GetAsyncEnumerator(); }
-public interface IDbAsyncEnumerator { object Current { get; } }
