+++ tests/X.PagedList.Tests/TestDbAsyncQueryProvider.cs
@@ -1,0 +1,148 @@
+﻿using System.Collections.Generic;
+using System.Linq;
+using System.Linq.Expressions;
+using System.Threading;
+using System.Threading.Tasks;
+
+namespace X.PagedList.Tests;
+
+public class TestDbAsyncQueryProvider { }
+internal class TestDbAsyncQueryProvider<TEntity> : IDbAsyncQueryProvider
+{
+    private readonly IQueryProvider _inner;
+
+    internal TestDbAsyncQueryProvider(IQueryProvider inner)
+    {
+        _inner = inner;
+    }
+
+    public IQueryable CreateQuery(Expression expression)
+    {
+        return new TestDbAsyncEnumerable<TEntity>(expression);
+    }
+
+    public IQueryable<TElement> CreateQuery<TElement>(Expression expression)
+    {
+        return new TestDbAsyncEnumerable<TElement>(expression);
+    }
+
+    public object Execute(Expression expression)
+    {
+        return _inner.Execute(expression);
+    }
+
+    public TResult Execute<TResult>(Expression expression)
+    {
+        return _inner.Execute<TResult>(expression);
+    }
+
+    public Task<object> ExecuteAsync(Expression expression, CancellationToken cancellationToken)
+    {
+        return Task.FromResult(Execute(expression));
+    }
+
+    public Task<TResult> ExecuteAsync<TResult>(Expression expression, CancellationToken cancellationToken)
+    {
+        return Task.FromResult(Execute<TResult>(expression));
+    }
+
+    internal class TestDbAsyncEnumerable<T> : EnumerableQuery<T>, IDbAsyncEnumerable<T>, IQueryable<T>
+    {
+        public TestDbAsyncEnumerable(IEnumerable<T> enumerable)
+            : base(enumerable)
+        { }
+
+        public TestDbAsyncEnumerable(Expression expression)
+            : base(expression)
+        { }
+
+        public IDbAsyncEnumerator<T> GetAsyncEnumerator()
+        {
+            return new TestDbAsyncEnumerator<T>(this.AsEnumerable().GetEnumerator());
+        }
+
+        IDbAsyncEnumerator IDbAsyncEnumerable.GetAsyncEnumerator()
+        {
+            return GetAsyncEnumerator();
+        }
+
+        IQueryProvider IQueryable.Provider
+        {
+            get { return new TestDbAsyncQueryProvider<T>(this); }
+        }
+    }
+
+    internal class TestDbAsyncEnumerator<T> : IDbAsyncEnumerator<T>
+    {
+        private readonly IEnumerator<T> _inner;
+
+        public TestDbAsyncEnumerator(IEnumerator<T> inner)
+        {
+            _inner = inner;
+        }
+
+        public void Dispose()
+        {
+            _inner.Dispose();
+        }
+
+        public Task<bool> MoveNextAsync(CancellationToken cancellationToken)
+        {
+            return Task.FromResult(_inner.MoveNext());
+        }
+
+        public T Current
+        {
+            get { return _inner.Current; }
+        }
+
+        object IDbAsyncEnumerator.Current
+        {
+            get { return Current; }
+        }
+    }
+
+    internal class AsyncEnumerableQuery<T> : EnumerableQuery<T>, IDbAsyncEnumerable<T>
+    {
+        public AsyncEnumerableQuery(IEnumerable<T> enumerable) : base(enumerable)
+        {
+        }
+
+        public AsyncEnumerableQuery(Expression expression) : base(expression)
+        {
+        }
+
+        public IDbAsyncEnumerator<T> GetAsyncEnumerator()
+        {
+            return new InMemoryDbAsyncEnumerator<T>(((IEnumerable<T>)this).GetEnumerator());
+        }
+
+        IDbAsyncEnumerator IDbAsyncEnumerable.GetAsyncEnumerator()
+        {
+            return GetAsyncEnumerator();
+        }
+
+        private class InMemoryDbAsyncEnumerator<T> : IDbAsyncEnumerator<T>
+        {
+            private readonly IEnumerator<T> _enumerator;
+
+            public InMemoryDbAsyncEnumerator(IEnumerator<T> enumerator)
+            {
+                _enumerator = enumerator;
+            }
+
+            public void Dispose()
+            {
+            }
+
+            public Task<bool> MoveNextAsync(CancellationToken cancellationToken)
+            {
+                return Task.FromResult(_enumerator.MoveNext());
+            }
+
+            public T Current => _enumerator.Current;
+
+            object IDbAsyncEnumerator.Current => Current;
+        }
+    }
+}
